document server and rename
