+++ server/server.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/kalbasit/signal-receiver/signalapireceiver"
+	"github.com/kalbasit/signal-receiver/receiver"
 )
 
 const usage = `
@@ -13,14 +13,30 @@ GET /receive/pop   => Return the oldest message
 GET /receive/flush => Return all messages
 `
 
+// Server represent the HTTP server that exposes the pop/flush routes
 type Server struct {
-	sarc *signalapireceiver.Client
+	sarc client
 }
 
-func New(sarc *signalapireceiver.Client) *Server {
+type client interface {
+	Pop() *receiver.Message
+	Flush() []receiver.Message
+}
+
+// New returns a new Server
+func New(sarc client) *Server {
 	return &Server{sarc: sarc}
 }
 
+// ServeHTTP implements the http.Handler interface
+//
+// /receive/pop
+//
+//	This returns status 200 and a receiver.Message or status 204 with no body
+//
+// /receive/flush
+//
+//	This returns status 200 and a list of receiver.Message
 func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodGet {
 		w.WriteHeader(http.StatusForbidden)
@@ -38,15 +54,20 @@ func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		if err := json.NewEncoder(w).Encode(msg); err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
 		}
-	} else if r.URL.Path == "/receive/flush" {
+
+		return
+	}
+
+	if r.URL.Path == "/receive/flush" {
 		msgs := s.sarc.Flush()
 		w.Header().Set("Content-Type", "application/json")
 		if err := json.NewEncoder(w).Encode(msgs); err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
 		}
-	} else {
-		w.Header().Set("Content-Type", "text/plain")
-		w.WriteHeader(http.StatusNotFound)
-		w.Write([]byte(fmt.Sprintf("ERROR! GET %s is not supported. The supported paths are below:", r.URL.Path) + usage))
+		return
 	}
+
+	w.Header().Set("Content-Type", "text/plain")
+	w.WriteHeader(http.StatusNotFound)
+	w.Write([]byte(fmt.Sprintf("ERROR! GET %s is not supported. The supported paths are below:", r.URL.Path) + usage))
 }
